add viability page and corntypepage
