--- assets/labels.txt
@@ -1,5 +1,0 @@
-Broken
-Discolored
-Pure
-Silkcut
-Undefined
