+++ assets/models/viabilityLabel.txt
@@ -1,0 +1,3 @@
+Less-Viable
+Non-Viable
+Viable
