+++ pubspec.yaml
@@ -79,12 +79,12 @@ flutter_launcher_icons:
 # The following section is specific to Flutter packages.
 flutter:
   assets:
-   - assets/labels.txt
    - assets/viability_labels.txt
-   - assets/model.tflite
    - assets/corn_viability.tflite
    - assets/models/Corn_types_float16.tflite
    - assets/models/labels.txt
+   - assets/models/viabilityLabel.txt
+   - assets/models/cornViability.tflite
    - assets/images/pic1.png
    - assets/images/handCorn.png
    - assets/images/cornSeedsDamo.png
